feat: Extract signature and message verification from CacaoVerifier
