+++ androidCore/sdk/src/main/kotlin/com/walletconnect/android/internal/common/model/params/CoreAuthParams.kt
@@ -3,7 +3,7 @@ package com.walletconnect.android.internal.common.model.params
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import com.walletconnect.android.internal.common.model.type.ClientParams
-import com.walletconnect.android.internal.common.cacao.Cacao
+import com.walletconnect.android.internal.common.signing.cacao.Cacao
 
 open class CoreAuthParams : ClientParams {
     @JsonClass(generateAdapter = true)
