+++ androidCore/sdk/src/main/kotlin/com/walletconnect/android/internal/common/signing/cacao/Cacao.kt
@@ -1,10 +1,10 @@
-package com.walletconnect.android.internal.common.cacao
+package com.walletconnect.android.internal.common.signing.cacao
 
 import androidx.annotation.Keep
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import com.walletconnect.android.cacao.SignatureInterface
-import com.walletconnect.android.internal.common.cacao.signature.Signature
+import com.walletconnect.android.internal.common.signing.signature.Signature
 
 @JsonClass(generateAdapter = true)
 data class Cacao(
