+++ androidCore/sdk/src/main/kotlin/com/walletconnect/android/internal/common/signing/cacao/CacaoType.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.android.internal.common.cacao
+package com.walletconnect.android.internal.common.signing.cacao
 
 enum class CacaoType(val header: String) {
     EIP4361("eip4361");
