+++ androidCore/sdk/src/main/kotlin/com/walletconnect/android/internal/common/signing/signature/SignatureInterface.kt
