+++ androidCore/sdk/src/main/kotlin/com/walletconnect/android/keyserver/domain/IdentitiesInteractor.kt
@@ -1,9 +1,9 @@
 package com.walletconnect.android.keyserver.domain
 
-import com.walletconnect.android.internal.common.cacao.Cacao
-import com.walletconnect.android.internal.common.cacao.CacaoType
-import com.walletconnect.android.internal.common.cacao.CacaoVerifier
-import com.walletconnect.android.internal.common.cacao.toCAIP122Message
+import com.walletconnect.android.internal.common.signing.cacao.Cacao
+import com.walletconnect.android.internal.common.signing.cacao.CacaoType
+import com.walletconnect.android.internal.common.signing.cacao.CacaoVerifier
+import com.walletconnect.android.internal.common.signing.cacao.toCAIP122Message
 import com.walletconnect.android.internal.common.crypto.kmr.KeyManagementRepository
 import com.walletconnect.android.internal.common.exception.*
 import com.walletconnect.android.internal.common.jwt.did.EncodeDidJwtPayloadUseCase
