+++ androidCore/sdk/src/test/kotlin/com/walletconnect/android/internal/common/cacao/IssuerTest.kt
@@ -1,5 +1,6 @@
 package com.walletconnect.android.internal.common.cacao
 
+import com.walletconnect.android.internal.common.signing.cacao.Issuer
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import kotlin.test.assertEquals
