+++ auth/sdk/src/main/kotlin/com/walletconnect/auth/client/mapper/ClientMapper.kt
@@ -2,8 +2,8 @@
 
 package com.walletconnect.auth.client.mapper
 
-import com.walletconnect.android.internal.common.cacao.Cacao
-import com.walletconnect.android.internal.common.cacao.CacaoType
+import com.walletconnect.android.internal.common.signing.cacao.Cacao
+import com.walletconnect.android.internal.common.signing.cacao.CacaoType
 import com.walletconnect.android.internal.common.model.ConnectionState
 import com.walletconnect.android.internal.common.model.SDKError
 import com.walletconnect.auth.client.Auth
