+++ auth/sdk/src/main/kotlin/com/walletconnect/auth/common/model/AuthResponse.kt
@@ -2,7 +2,7 @@
 
 package com.walletconnect.auth.common.model
 
-import com.walletconnect.android.internal.common.cacao.Cacao
+import com.walletconnect.android.internal.common.signing.cacao.Cacao
 
 internal sealed class AuthResponse {
     data class Result(val cacao: Cacao) : AuthResponse()
