+++ auth/sdk/src/main/kotlin/com/walletconnect/auth/di/EngineModule.kt
@@ -1,6 +1,6 @@
 package com.walletconnect.auth.di
 
-import com.walletconnect.android.internal.common.cacao.CacaoVerifier
+import com.walletconnect.android.internal.common.signing.cacao.CacaoVerifier
 import com.walletconnect.android.internal.common.di.AndroidCommonDITags
 import com.walletconnect.auth.engine.domain.AuthEngine
 import com.walletconnect.auth.json_rpc.domain.GetPendingJsonRpcHistoryEntriesUseCase
