+++ push/sdk/src/main/kotlin/com/walletconnect/push/wallet/engine/PushWalletEngine.kt
@@ -3,7 +3,7 @@
 package com.walletconnect.push.wallet.engine
 
 import com.walletconnect.android.CoreClient
-import com.walletconnect.android.internal.common.cacao.Cacao
+import com.walletconnect.android.internal.common.signing.cacao.Cacao
 import com.walletconnect.android.internal.common.crypto.codec.Codec
 import com.walletconnect.android.internal.common.crypto.kmr.KeyManagementRepository
 import com.walletconnect.android.internal.common.crypto.sha256
